cleanup doc output
