--- docs/MockValidator.html
@@ -1,148 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-<head>
-  <title>MockValidator.ts</title>
-  <meta http-equiv="content-type" content="text/html; charset=UTF-8">
-  <link rel="stylesheet" media="all" href="public/stylesheets/normalize.css" />
-  <link rel="stylesheet" media="all" href="docco.css" />
-</head>
-<body>
-  <div class="container">
-    <div class="page">
-
-      <div class="header">
-        
-          <h1>MockValidator.ts</h1>
-        
-
-        
-          <div class="toc">
-            <h3>Table of Contents</h3>
-            <ol>
-              
-                
-                <li>
-                  <a class="source" href="MockValidator.html">
-                    MockValidator.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="dereference.test.html">
-                    dereference.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="get.test.html">
-                    get.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="isPointer.test.html">
-                    isPointer.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="mockResolve.html">
-                    mockResolve.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="set.test.html">
-                    set.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="index.html">
-                    index.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="json.d.html">
-                    json.d.ts
-                  </a>
-                </li>
-              
-            </ol>
-          </div>
-        
-      </div>
-
-      
-        
-        
-        
-          <div class='highlight'><pre><span class="hljs-keyword">import</span> * <span class="hljs-keyword">as</span> Ajv <span class="hljs-keyword">from</span> <span class="hljs-string">'ajv'</span>;
-
-<span class="hljs-keyword">import</span> * <span class="hljs-keyword">as</span> address <span class="hljs-keyword">from</span> <span class="hljs-string">'./fixture/address+v1.schema.json'</span>;
-<span class="hljs-keyword">import</span> * <span class="hljs-keyword">as</span> addressOverride <span class="hljs-keyword">from</span> <span class="hljs-string">'./fixture/address-override+v1.schema.json'</span>;
-<span class="hljs-keyword">import</span> * <span class="hljs-keyword">as</span> circular <span class="hljs-keyword">from</span> <span class="hljs-string">'./fixture/circular.schema.json'</span>;
-<span class="hljs-keyword">import</span> * <span class="hljs-keyword">as</span> conditional <span class="hljs-keyword">from</span> <span class="hljs-string">'./fixture/conditional.schema.json'</span>;
-<span class="hljs-keyword">import</span> * <span class="hljs-keyword">as</span> credentials <span class="hljs-keyword">from</span> <span class="hljs-string">'./fixture/credentials+v1.schema.json'</span>;
-<span class="hljs-keyword">import</span> * <span class="hljs-keyword">as</span> editPerson <span class="hljs-keyword">from</span> <span class="hljs-string">'./fixture/edit-person+v1.schema.json'</span>;
-<span class="hljs-keyword">import</span> * <span class="hljs-keyword">as</span> person <span class="hljs-keyword">from</span> <span class="hljs-string">'./fixture/person+v1.schema.json'</span>;
-<span class="hljs-keyword">import</span> * <span class="hljs-keyword">as</span> profile <span class="hljs-keyword">from</span> <span class="hljs-string">'./fixture/profile+v1.schema.json'</span>;
-
-<span class="hljs-comment">/**
- * A validation helper for our test schema.
- */</span>
-<span class="hljs-keyword">export</span> <span class="hljs-keyword">class</span> MockValidator {
-  <span class="hljs-keyword">public</span> schema: object[] = [];
-  <span class="hljs-keyword">private</span> ajv;
-
-  <span class="hljs-keyword">constructor</span>(<span class="hljs-params">config: object = {}</span>) {
-    <span class="hljs-keyword">const</span> schema = [
-      address,
-      addressOverride,
-      circular,
-      conditional,
-      credentials,
-      editPerson,
-      person,
-      profile,
-    ];
-    <span class="hljs-keyword">const</span> cnf = {
-      extendRefs: <span class="hljs-literal">true</span>,
-      allErrors: <span class="hljs-literal">true</span>,
-      ...config,
-    };
-    <span class="hljs-keyword">this</span>.ajv = <span class="hljs-keyword">new</span> Ajv(cnf);
-    <span class="hljs-keyword">this</span>.load(schema);
-  }
-
-  load(schema: object[]) {
-    schema.map(<span class="hljs-function">(<span class="hljs-params">scm</span>) =&gt;</span> {
-      <span class="hljs-keyword">this</span>.ajv.addSchema(scm);
-      <span class="hljs-keyword">this</span>.schema.push(scm);
-    });
-
-    <span class="hljs-keyword">return</span> <span class="hljs-keyword">this</span>;
-  }
-
-  getSchema(id: <span class="hljs-built_in">string</span>): object | <span class="hljs-literal">undefined</span> {
-    <span class="hljs-keyword">const</span> lookup = <span class="hljs-keyword">this</span>.ajv.getSchema(id);
-
-    <span class="hljs-keyword">return</span> !lookup ? <span class="hljs-literal">undefined</span> : lookup.schema;
-  }
-}</pre></div>
-        
-      
-      <div class="fleur">h</div>
-    </div>
-  </div>
-</body>
-</html>
