--- docs/get.test.html
@@ -1,159 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-<head>
-  <title>get.test.ts</title>
-  <meta http-equiv="content-type" content="text/html; charset=UTF-8">
-  <link rel="stylesheet" media="all" href="public/stylesheets/normalize.css" />
-  <link rel="stylesheet" media="all" href="docco.css" />
-</head>
-<body>
-  <div class="container">
-    <div class="page">
-
-      <div class="header">
-        
-          <h1>get.test.ts</h1>
-        
-
-        
-          <div class="toc">
-            <h3>Table of Contents</h3>
-            <ol>
-              
-                
-                <li>
-                  <a class="source" href="MockValidator.html">
-                    MockValidator.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="dereference.test.html">
-                    dereference.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="get.test.html">
-                    get.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="isPointer.test.html">
-                    isPointer.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="mockResolve.html">
-                    mockResolve.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="set.test.html">
-                    set.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="index.html">
-                    index.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="json.d.html">
-                    json.d.ts
-                  </a>
-                </li>
-              
-            </ol>
-          </div>
-        
-      </div>
-
-      
-        
-        
-        
-          <div class='highlight'><pre><span class="hljs-keyword">import</span> { expect } <span class="hljs-keyword">from</span> <span class="hljs-string">'chai'</span>;
-<span class="hljs-keyword">import</span> { <span class="hljs-keyword">get</span> } <span class="hljs-keyword">from</span> <span class="hljs-string">'./../index'</span>;
-
-describe(<span class="hljs-string">'@jdw/jst/get'</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-  <span class="hljs-keyword">const</span> fixture = {
-    <span class="hljs-string">'foo'</span>: {
-      bar: <span class="hljs-number">20</span>,
-    },
-    <span class="hljs-string">'/'</span>: {
-      <span class="hljs-string">'~'</span>: <span class="hljs-number">400</span>,
-      <span class="hljs-string">'bar'</span>: {
-        <span class="hljs-string">'~'</span>: <span class="hljs-number">10</span>,
-      },
-    },
-    <span class="hljs-string">'arr'</span>: [{
-      arr: [{
-        foo: <span class="hljs-number">100</span>,
-      }],
-    }, <span class="hljs-number">30</span>],
-  };
-
-  it(<span class="hljs-string">'should resolve valid JSON pointers'</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-    <span class="hljs-keyword">const</span> cases = [
-      [<span class="hljs-string">'/foo/bar'</span>, <span class="hljs-number">20</span>],
-      [<span class="hljs-string">'#/foo/bar'</span>, <span class="hljs-number">20</span>],
-      [<span class="hljs-string">''</span>, fixture],
-    ];
-    cases.forEach(<span class="hljs-function">(<span class="hljs-params">t</span>) =&gt;</span> {
-      expect(<span class="hljs-keyword">get</span>(fixture, t[<span class="hljs-number">0</span>].toString())).to.eq(t[<span class="hljs-number">1</span>]);
-    });
-  });
-
-  it(<span class="hljs-string">'can resolve the encoded values of the / and ~ symbols correctly'</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-    <span class="hljs-keyword">const</span> cases = [
-      [<span class="hljs-string">'#/~1/~0'</span>, <span class="hljs-number">400</span>],
-      [<span class="hljs-string">'#/~1/bar/~0'</span>, <span class="hljs-number">10</span>],
-    ];
-    cases.forEach(<span class="hljs-function">(<span class="hljs-params">t</span>) =&gt;</span> {
-      expect(<span class="hljs-keyword">get</span>(fixture, t[<span class="hljs-number">0</span>].toString())).to.eq(t[<span class="hljs-number">1</span>]);
-    });
-  });
-
-  it(<span class="hljs-string">'can resolve JSON pointers that reference array indexes correctly'</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-    <span class="hljs-keyword">const</span> cases = [
-      [<span class="hljs-string">'#/arr/0/arr/0/foo'</span>, <span class="hljs-number">100</span>],
-      [<span class="hljs-string">'#/arr/1'</span>, <span class="hljs-number">30</span>],
-    ];
-    cases.forEach(<span class="hljs-function">(<span class="hljs-params">t</span>) =&gt;</span> {
-      expect(<span class="hljs-keyword">get</span>(fixture, t[<span class="hljs-number">0</span>].toString())).to.eq(t[<span class="hljs-number">1</span>]);
-    });
-  });
-
-  it(<span class="hljs-string">'should throw an error up-front when providing an invalid json pointer'</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-    <span class="hljs-keyword">const</span> cases = [
-      <span class="hljs-string">'abcd'</span>,
-      <span class="hljs-string">'~#/foo/bar'</span>,
-      <span class="hljs-string">'foo/bar'</span>,
-    ];
-
-    cases.forEach(<span class="hljs-function">(<span class="hljs-params">t</span>) =&gt;</span> {
-      expect(<span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> <span class="hljs-keyword">get</span>(fixture, t)).to.throw();
-    });
-  });
-});</pre></div>
-        
-      
-      <div class="fleur">h</div>
-    </div>
-  </div>
-</body>
-</html>
