--- docs/isPointer.test.html
@@ -1,116 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-<head>
-  <title>isPointer.test.ts</title>
-  <meta http-equiv="content-type" content="text/html; charset=UTF-8">
-  <link rel="stylesheet" media="all" href="public/stylesheets/normalize.css" />
-  <link rel="stylesheet" media="all" href="docco.css" />
-</head>
-<body>
-  <div class="container">
-    <div class="page">
-
-      <div class="header">
-        
-          <h1>isPointer.test.ts</h1>
-        
-
-        
-          <div class="toc">
-            <h3>Table of Contents</h3>
-            <ol>
-              
-                
-                <li>
-                  <a class="source" href="MockValidator.html">
-                    MockValidator.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="dereference.test.html">
-                    dereference.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="get.test.html">
-                    get.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="isPointer.test.html">
-                    isPointer.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="mockResolve.html">
-                    mockResolve.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="set.test.html">
-                    set.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="index.html">
-                    index.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="json.d.html">
-                    json.d.ts
-                  </a>
-                </li>
-              
-            </ol>
-          </div>
-        
-      </div>
-
-      
-        
-        
-        
-          <div class='highlight'><pre><span class="hljs-keyword">import</span> { expect } <span class="hljs-keyword">from</span> <span class="hljs-string">'chai'</span>;
-<span class="hljs-keyword">import</span> { isPointer } <span class="hljs-keyword">from</span> <span class="hljs-string">'./../index'</span>;
-
-describe(<span class="hljs-string">'@jdw/jst/isPointer'</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-  <span class="hljs-keyword">const</span> cases = {
-    success: [<span class="hljs-string">'#'</span>, <span class="hljs-string">'/'</span>, <span class="hljs-string">''</span>, <span class="hljs-string">'#/foo/bar/baz'</span>, <span class="hljs-string">'/bar/baz'</span>],
-    failure: [<span class="hljs-string">'yolo'</span>, <span class="hljs-number">123</span>, {}, <span class="hljs-literal">false</span>, <span class="hljs-literal">true</span>, <span class="hljs-literal">null</span>, <span class="hljs-literal">undefined</span>], <span class="hljs-comment">//  '#adasd']</span>
-  };
-
-  cases.success.forEach(<span class="hljs-function">(<span class="hljs-params">testCase</span>) =&gt;</span> {
-    it(<span class="hljs-string">`can recognize <span class="hljs-subst">${JSON.stringify(testCase)}</span> as a pointer`</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-      expect(isPointer(testCase)).to.eq(<span class="hljs-literal">true</span>);
-    });
-  });
-
-  cases.failure.forEach(<span class="hljs-function">(<span class="hljs-params">testCase</span>) =&gt;</span> {
-    it(<span class="hljs-string">`can recognize <span class="hljs-subst">${JSON.stringify(testCase)}</span> is NOT a pointer`</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-      expect(isPointer(testCase)).to.eq(<span class="hljs-literal">false</span>);
-    });
-  });
-});</pre></div>
-        
-      
-      <div class="fleur">h</div>
-    </div>
-  </div>
-</body>
-</html>
