--- docs/json.d.html
@@ -1,99 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-<head>
-  <title>json.d.ts</title>
-  <meta http-equiv="content-type" content="text/html; charset=UTF-8">
-  <link rel="stylesheet" media="all" href="public/stylesheets/normalize.css" />
-  <link rel="stylesheet" media="all" href="docco.css" />
-</head>
-<body>
-  <div class="container">
-    <div class="page">
-
-      <div class="header">
-        
-          <h1>json.d.ts</h1>
-        
-
-        
-          <div class="toc">
-            <h3>Table of Contents</h3>
-            <ol>
-              
-                
-                <li>
-                  <a class="source" href="MockValidator.html">
-                    MockValidator.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="dereference.test.html">
-                    dereference.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="get.test.html">
-                    get.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="isPointer.test.html">
-                    isPointer.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="mockResolve.html">
-                    mockResolve.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="set.test.html">
-                    set.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="index.html">
-                    index.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="json.d.html">
-                    json.d.ts
-                  </a>
-                </li>
-              
-            </ol>
-          </div>
-        
-      </div>
-
-      
-        
-        
-        
-          <div class='highlight'><pre><span class="hljs-keyword">declare</span> <span class="hljs-keyword">module</span> "*.json" {
-  <span class="hljs-keyword">const</span> value: <span class="hljs-built_in">any</span>;
-  <span class="hljs-keyword">export</span> <span class="hljs-keyword">default</span> value;
-}</pre></div>
-        
-      
-      <div class="fleur">h</div>
-    </div>
-  </div>
-</body>
-</html>
