--- docs/mockResolve.html
@@ -1,114 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-<head>
-  <title>mockResolve.ts</title>
-  <meta http-equiv="content-type" content="text/html; charset=UTF-8">
-  <link rel="stylesheet" media="all" href="public/stylesheets/normalize.css" />
-  <link rel="stylesheet" media="all" href="docco.css" />
-</head>
-<body>
-  <div class="container">
-    <div class="page">
-
-      <div class="header">
-        
-          <h1>mockResolve.ts</h1>
-        
-
-        
-          <div class="toc">
-            <h3>Table of Contents</h3>
-            <ol>
-              
-                
-                <li>
-                  <a class="source" href="MockValidator.html">
-                    MockValidator.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="dereference.test.html">
-                    dereference.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="get.test.html">
-                    get.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="isPointer.test.html">
-                    isPointer.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="mockResolve.html">
-                    mockResolve.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="set.test.html">
-                    set.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="index.html">
-                    index.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="json.d.html">
-                    json.d.ts
-                  </a>
-                </li>
-              
-            </ol>
-          </div>
-        
-      </div>
-
-      
-        
-        
-        
-          <div class='highlight'><pre><span class="hljs-keyword">import</span> { MockValidator <span class="hljs-keyword">as</span> Validator } <span class="hljs-keyword">from</span> <span class="hljs-string">'./MockValidator'</span>;
-
-<span class="hljs-keyword">const</span> validator = <span class="hljs-keyword">new</span> Validator();
-
-<span class="hljs-comment">/**
- * A resolve function must simply take a schema id as an argument and return
- * that schema as an object literal or throw an error if it can't find it.
- */</span>
-<span class="hljs-keyword">const</span> mockResolve = <span class="hljs-function">(<span class="hljs-params">id</span>) =&gt;</span> {
-  <span class="hljs-keyword">const</span> result = validator.getSchema(id);
-
-  <span class="hljs-keyword">if</span> (!result) {
-    <span class="hljs-keyword">throw</span> <span class="hljs-keyword">new</span> <span class="hljs-built_in">Error</span>(<span class="hljs-string">`could not resolve schema with id: <span class="hljs-subst">${id}</span>`</span>);
-  }
-
-  <span class="hljs-keyword">return</span> result;
-};
-
-<span class="hljs-keyword">export</span> { mockResolve };</pre></div>
-        
-      
-      <div class="fleur">h</div>
-    </div>
-  </div>
-</body>
-</html>
