--- docs/set.test.html
@@ -1,133 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-<head>
-  <title>set.test.ts</title>
-  <meta http-equiv="content-type" content="text/html; charset=UTF-8">
-  <link rel="stylesheet" media="all" href="public/stylesheets/normalize.css" />
-  <link rel="stylesheet" media="all" href="docco.css" />
-</head>
-<body>
-  <div class="container">
-    <div class="page">
-
-      <div class="header">
-        
-          <h1>set.test.ts</h1>
-        
-
-        
-          <div class="toc">
-            <h3>Table of Contents</h3>
-            <ol>
-              
-                
-                <li>
-                  <a class="source" href="MockValidator.html">
-                    MockValidator.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="dereference.test.html">
-                    dereference.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="get.test.html">
-                    get.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="isPointer.test.html">
-                    isPointer.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="mockResolve.html">
-                    mockResolve.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="set.test.html">
-                    set.test.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="index.html">
-                    index.ts
-                  </a>
-                </li>
-              
-                
-                <li>
-                  <a class="source" href="json.d.html">
-                    json.d.ts
-                  </a>
-                </li>
-              
-            </ol>
-          </div>
-        
-      </div>
-
-      
-        
-        
-        
-          <div class='highlight'><pre><span class="hljs-keyword">import</span> { expect } <span class="hljs-keyword">from</span> <span class="hljs-string">'chai'</span>;
-<span class="hljs-keyword">import</span> { <span class="hljs-keyword">set</span> } <span class="hljs-keyword">from</span> <span class="hljs-string">'./../index'</span>;
-
-describe(<span class="hljs-string">'@jdw/jst/set'</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-  <span class="hljs-keyword">const</span> fixture = {
-    array: [<span class="hljs-number">1</span>, <span class="hljs-number">2</span>, <span class="hljs-literal">null</span>],
-    object: {
-      foo: <span class="hljs-literal">null</span>,
-      array: [{ foo: <span class="hljs-literal">null</span> }, <span class="hljs-literal">null</span>],
-    },
-    foo: <span class="hljs-literal">null</span>,
-  };
-
-  it(<span class="hljs-string">'can set an property on an object'</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-    <span class="hljs-keyword">set</span>(fixture, <span class="hljs-string">'#/foo'</span>, <span class="hljs-string">'bar'</span>);
-    expect(fixture.foo).to.eq(<span class="hljs-string">'bar'</span>);
-  });
-
-  it(<span class="hljs-string">'can set an index on an array'</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-    <span class="hljs-keyword">set</span>(fixture, <span class="hljs-string">'#/array/2'</span>, <span class="hljs-number">3</span>);
-    expect(fixture.array[<span class="hljs-number">2</span>]).to.eq(<span class="hljs-number">3</span>);
-  });
-
-  it(<span class="hljs-string">'can set an property on a nested object'</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-    <span class="hljs-keyword">set</span>(fixture, <span class="hljs-string">'#/object/foo'</span>, <span class="hljs-string">'bar'</span>);
-    expect(fixture.object.foo).to.eq(<span class="hljs-string">'bar'</span>);
-  });
-
-  it(<span class="hljs-string">'can set an property on a nested array'</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-    <span class="hljs-keyword">set</span>(fixture, <span class="hljs-string">'#/object/array/1'</span>, <span class="hljs-string">'bar'</span>);
-    expect(fixture.object.array[<span class="hljs-number">1</span>]).to.eq(<span class="hljs-string">'bar'</span>);
-  });
-
-  it(<span class="hljs-string">'can set an property on an object in an array'</span>, <span class="hljs-function"><span class="hljs-params">()</span> =&gt;</span> {
-    <span class="hljs-keyword">set</span>(fixture, <span class="hljs-string">'#/object/array/0/foo'</span>, <span class="hljs-string">'bar'</span>);
-    expect(fixture.object.array[<span class="hljs-number">0</span>].foo).to.eq(<span class="hljs-string">'bar'</span>);
-  });
-});</pre></div>
-        
-      
-      <div class="fleur">h</div>
-    </div>
-  </div>
-</body>
-</html>
