+++ src/__benchmarks__/index.benchmark.ts
