+++ src/__tests__/lib/index.ts
@@ -1,0 +1,7 @@
+import { Validator } from './Validator';
+import { resolve } from './resolve';
+
+export {
+  Validator,
+  resolve,
+};
